// ph-healthcare-frontent/src/generated/prisma/enums.ts
export const RecordStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type RecordStatus = (typeof RecordStatus)[keyof typeof RecordStatus]
// ...
export const LabRequestStatus = {
  REQUESTED: 'REQUESTED',
  SAMPLE_COLLECTED: 'SAMPLE_COLLECTED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type LabRequestStatus = (typeof LabRequestStatus)[keyof typeof LabRequestStatus]


export const LabPriority = {
  ROUTINE: 'ROUTINE',
  URGENT: 'URGENT',
  STAT: 'STAT'
} as const

export type LabPriority = (typeof LabPriority)[keyof typeof LabPriority]